refactor: rename controllers to Api and update package structure for consistency
