+++ src/main/java/org/thingai/app/api/MatchControlApi.java
@@ -1,10 +1,10 @@
-package org.thingai.app.controller;
+package org.thingai.app.api;
 
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/scorekeeper")
-public class ScorekeeperController {
+public class MatchControlApi {
 
 
 }
