+++ src/main/java/org/thingai/app/api/RefereeApi.java
@@ -1,10 +1,10 @@
-package org.thingai.app.controller;
+package org.thingai.app.api;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
 @RequestMapping("/api/ref")
-public class RefereeController {
+public class RefereeApi {
 
 }
